Move Controls to atoms
